+++ src/test/java/test/TestSDAXPath.java
@@ -78,7 +78,7 @@ public class TestSDAXPath {
 		t.so("S31", "name()", alice, "contact");
 		t.so("S32", "name()", doc, ""); // returns empty string
 		
-		System.out.print("\n            string ");
+		System.out.print("\n	string ");
 
 		t.so("S33", "string(/addressbook/contact)", doc, "1");
 		t.so("S34", "string()", addressbook, "");
@@ -123,7 +123,7 @@ public class TestSDAXPath {
 		t.so("S62", "false()", doc, "false");		
 		t.so("S63", "lang(addressbook)", doc, "false");	
 		
-		System.out.print("\n            number ");
+		System.out.print("\n	number ");
 		
 		t.so("S64", "number(/addressbook/contact)", doc, "1.0");
 		t.so("S65", "number()", bob, "2.0");
@@ -133,7 +133,7 @@ public class TestSDAXPath {
 		t.so("S69", "round(1.5)", doc, "2.0");
 		t.so("S70", "((1+1)*2mod3)div2-1", doc, "-0.5");
 		
-		System.out.print(" jaxen ");
+		System.out.print("jaxen ");
 
 		t.so("S71", "ends-with(/addressbook/contact/firstname,'e')", doc, "true");
 		t.so("S72", "ends-with(firstname,'b')", bob, "true");
@@ -148,7 +148,7 @@ public class TestSDAXPath {
 		
 		t.so("S79", "document('"+ file + "')", addressbook, "["+doc.toString()+"]");
 		
-		System.out.print(" sdt ");
+		System.out.print("sdt ");
 		
 		t.so("S80", "fn:string-join(/addressbook/contact/phonenumber)", doc, "06-1111111106-2222222206-3333333306-44444444");
 		t.so("S81", "fn:string-join(contact | contact/firstname,':')", addressbook, "1:Alice:2:Bob");
@@ -161,21 +161,32 @@ public class TestSDAXPath {
 		t.so("S86", "sdt:right(/addressbook/contact[2]/firstname,2)", doc, "ob");
 		t.so("S87", "sdt:right(/addressbook/contact[2]/firstname,4)", doc, "Bob");
 		
-		t.so("S88", "sdt:compare-number(1,3)", doc, "-1.0");
-		t.so("S89", "sdt:compare-number(3,'3')", doc, "0.0");
-		t.so("S90", "sdt:compare-number('6','4')", doc, "1.0");
-		t.so("S91", "sdt:compare-number('a',1)", doc, "1.0");
-		t.so("S92", "sdt:compare-number('a','b')", doc, "0.0");
-		t.so("S91", "sdt:compare-number('a',1,true())", doc, "-1.0");
-		t.so("S91", "sdt:compare-number('a',1,false())", doc, "1.0");
-		t.so("S92", "sdt:compare-number('a','b',true())", doc, "0.0");
-		t.so("S93", "sdt:compare-number('a','b',false())", doc, "0.0");
+		System.out.print("\n	");
 		
-		t.so("S94", "sdt:compare-string('a','A')", doc, "-1.0");
-		t.so("S95", "sdt:compare-string(3,'3')", doc, "0.0");
-		t.so("S96", "sdt:compare-string('b','A')", doc, "1.0");
-		t.so("S97", "sdt:compare-string('Ångström','Zulu','en')", doc, "-1.0");
-		t.so("S98", "sdt:compare-string('Ångström','Zulu','sv')", doc, "1.0");
+		t.so("S90", "sdt:compare-number(1,3)", doc, "-1.0");
+		t.so("S91", "sdt:compare-number(3,'3')", doc, "0.0");
+		t.so("S92", "sdt:compare-number('6','4')", doc, "1.0");
+		t.so("S93", "sdt:compare-number('a',1)", doc, "1.0");
+		t.so("S94", "sdt:compare-number('a','b')", doc, "0.0");
+		t.so("S95", "sdt:compare-number('a',1,true())", doc, "-1.0");
+		t.so("S96", "sdt:compare-number('a',1,false())", doc, "1.0");
+		t.so("S97", "sdt:compare-number('a','b',true())", doc, "0.0");
+		t.so("S98", "sdt:compare-number('a','b',false())", doc, "0.0");
+		
+		t.so("S100", "sdt:compare-string('a','A')", doc, "-1.0");
+		t.so("S101", "sdt:compare-string(3,'3')", doc, "0.0");
+		t.so("S102", "sdt:compare-string('b','A')", doc, "1.0");
+		t.so("S103", "sdt:compare-string('Ångström','Zulu','en')", doc, "-1.0");
+		t.so("S104", "sdt:compare-string('Ångström','Zulu','sv')", doc, "1.0");
+		
+		t.so("S110", "sdt:tokenize('')", doc, "[]");
+		t.so("S111", "sdt:tokenize('abc')", doc, "abc");
+		t.so("S112", "sdt:tokenize('abc','')", doc, "[a, b, c]");
+		t.so("S113", "sdt:tokenize(' a  b   c    ')", doc, "[a, b, c]");
+		t.so("S114", "sdt:tokenize('127.0.0.1:80','[\\.:]')", doc, "[127, 0, 0, 1, 80]");
+		t.so("S115", "sdt:tokenize('1;2;;3;',';')", doc, "[1, 2, 3]");
+		t.so("S116", "sdt:tokenize('1; 2; ; 3; ','; ',true())", doc, "[1, 2, , 3, ]");
+		
 	}
 
 }