+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@
 
 This is work in progress.
 
+- Added tokenize() function.
+- Fixed issue in sort comparator and statement context.
 - Closed issue #12 (possible to overwrite parameter).
 
 ## Compatibility